refactor(MutatorFactory) Move MutatorFactory to the API folder
