+++ src/api/mutant.ts
@@ -1,2 +1,3 @@
 export {default as StrykerNode} from './mutant/StrykerNode';
 export {default as Mutator} from './mutant/Mutator';
+export {default as MutatorFactory} from './mutant/MutatorFactory';
